webapp/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
flask-login==0.6.2
flask-session==0.5.0
mysql-connector-python==8.1.0
Flask-LoginManager==1.1.6
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
pyotp==2.9.0
